Implemented GetMentalHealthScoreForUser(), which returns a single summary number representing the mental health of the user. It does this by taking a mean of all the entries the user has entered.

+++ pkg/database/dataInterfaces.go
@@ -10,7 +10,7 @@ import (
 // Repository - interface for a data provider that interfaces between the database backend and the grpc server
 // enables the repository pattern so that we can swap out the database backend easily
 type Repository interface {
-	GetOverallScore(ctx context.Context, userID int64) (int, error)
+	GetOverallScore(ctx context.Context, userID int64) (float64, error)
 	GetAllMentalHealthLogs(ctx context.Context, userID int64) ([]*pbhealth.MentalHealthLog, error)
 	GetAllMentalHealthLogsByDate(ctx context.Context, userID int64, date *pbcommon.Date) ([]*pbhealth.MentalHealthLog, error)
 	AddMentalHealthLog(ctx context.Context, healthLog *pbhealth.MentalHealthLog) (string, error)
